Move test from emit to semantics
